+++ R/plume-package.R
@@ -3,13 +3,14 @@
 #' @keywords internal
 #' @importFrom dplyr distinct filter arrange select dense_rank
 #' @importFrom dplyr mutate summarise across
-#' @importFrom dplyr if_else
+#' @importFrom dplyr if_else if_any
 #' @importFrom tidyr unite drop_na pivot_longer nest unnest expand_grid
 #' @importFrom tidyselect all_of any_of starts_with
 #' @importFrom tibble tibble as_tibble_row as_tibble rowid_to_column
-#' @importFrom purrr set_names partial list_assign reduce
+#' @importFrom purrr set_names partial negate reduce
 #' @importFrom purrr is_empty
-#' @importFrom purrr map map_vec imap iwalk walk2 list_rbind list_flatten
+#' @importFrom purrr map map_vec map_chr imap iwalk walk2
+#' @importFrom purrr list_rbind list_flatten list_transpose list_assign
 #' @importFrom readr read_file write_lines
 #' @importFrom rlang %||% := abort
 #' @importFrom rlang expr exprs parse_expr sym syms