+++ NEWS.md
@@ -1,5 +1,9 @@
 # plume (development version)
 
-## Bug fixes
+## Minor improvements and bug fixes
+
+* `get_contact_details()` now drops corresponding authors with no contact details.
+
+* `get_contact_details()` has been reworked to bind any combination of contact details properly.
 
 * `to_yaml()` now outputs verbatim `true`/`false` (#1).